refactor common test code into separate module
