+++ server/tests/post-vehicle-positions.js
@@ -19,7 +19,7 @@ aren't identical, the test fails.
 const crypto = require('crypto').webcrypto
 const GtfsRealtimeBindings = require('gtfs-realtime-bindings');
 const util = require('../app-engine/static/gtfs-rt-util');
-const fetch = require('node-fetch');
+const testutil = require('./test-util');
 
 const VEHICLE_IDS = [
     'pr-test-vehicle-id-1', 'pr-test-vehicle-id-2',
@@ -28,64 +28,6 @@ const VEHICLE_IDS = [
     'pr-test-vehicle-id-5', 'pr-test-vehicle-id-5',
     'pr-test-vehicle-id-6', 'pr-test-vehicle-id-7'
 ];
-
-function getBaseName(s) {
-    const index = s.lastIndexOf('/');
-
-    if (index < 0) return s;
-    else return s.substring(index + 1);
-}
-
-async function getResponseBody(url) {
-    const requestSettings = {
-        method: 'GET'
-    };
-
-    const response = await fetch(url, requestSettings);
-
-    const blob = await response.blob();
-    //util.log(`- blob: ${blob}`);
-
-    const arrayBuf = await blob.arrayBuffer();
-    //util.log(`- arrayBuf.byteLength: ${arrayBuf.byteLength}`);
-
-    const body = Buffer.from(arrayBuf);
-    //util.log(`- body: ${JSON.stringify(body)}`);
-
-    return body;
-}
-
-async function getSignatureKey() {
-    const base64 = process.env.PR_TEST_ID_ECDSA;
-
-    if (!base64) throw 'unset environment variable $PR_TEST_ID_ECDSA';
-
-    const key = atob(base64);
-    util.log("- key.length: " + key.length);
-
-    const binaryDer = util.str2ab(key);
-
-    return await crypto.subtle.importKey(
-        "pkcs8",
-        binaryDer,
-        {
-            name: "ECDSA",
-            namedCurve: "P-256"
-        },
-        false,
-        ["sign"]
-    );
-}
-
-function sleep(ms) {
-    return new Promise((resolve) => {
-        setTimeout(resolve, ms);
-    });
-}
-
-function getEpochSeconds() {
-    return Math.floor(Date.now() / 1000);
-}
 
 async function postUpdates(signatureKey, url) {
     let data = {
@@ -112,11 +54,11 @@ async function postUpdates(signatureKey, url) {
         data['vehicle-id'] = vid;
         util.log(`-- vid: ${vid}`);
 
-        data['timestamp'] = getEpochSeconds();
+        data['timestamp'] = testutil.getEpochSeconds();
         util.log(`++ timestamp: ${data.timestamp}`);
 
-        util.signAndPost(data, signatureKey, url);
-        await sleep(1000);
+        await util.signAndPost(data, signatureKey, url);
+        await testutil.sleep(1000);
     }
 }
 
@@ -126,22 +68,23 @@ async function test(url) {
 
     process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0";
 
-    const signatureKey = await getSignatureKey();
+    const signatureKey = await testutil.getSignatureKey();
     util.log(`- signatureKey: ${JSON.stringify(signatureKey)}`);
 
     await postUpdates(signatureKey, url + '/new-pos-sig');
 
-    console.log('sleeping for a few seconds before accessing vehicle position feed...');
+    const sleepCount = 30;
+    console.log(`sleeping for ${sleepCount} seconds before accessing vehicle position feed...`);
 
-    for (let i=0; i<5; i++) {
-        await sleep(1000);
+    for (let i=0; i<sleepCount; i++) {
+        await testutil.sleep(1000);
         console.log('.');
     }
 
     console.log('done');
 
-    const now = getEpochSeconds();
-    const body = await getResponseBody(url + '/vehicle-positions.pb?agency=test');
+    const now = testutil.getEpochSeconds();
+    const body = await testutil.getResponseBody(url + '/vehicle-positions.pb?agency=test');
     const feed = GtfsRealtimeBindings.transit_realtime.FeedMessage.decode(body);
 
     let expectedVIDS = [];
